move enums file to constants folder
