+++ lib/db/DB.js
@@ -1,7 +1,7 @@
 const Sequelize = require('sequelize');
 const assert = require('assert');
 const Logger = require('../Logger');
-const enums = require('../enums');
+const enums = require('../constants/enums');
 
 const { DataTypes } = Sequelize;
